requirements.txt
astroid==2.3.1
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
dataclasses==0.6
defusedxml==0.6.0
Django==2.2.5
django-allauth==0.40.0
django-ckeditor==5.7.1
django-js-asset==1.2.2
django-widget-tweaks==1.4.5
djongo==1.2.36
dnspython==1.16.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
mock==3.0.5
oauthlib==3.1.0
Pillow==6.2.0
pylint==2.4.2
pymongo==3.9.0
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.2.4
typed-ast==1.4.0
urllib3==1.25.6
wrapt==1.11.2
